+++ app/Livewire/Budgetplan/Topic.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire\Budgetplan;
+
+use Livewire\Component;
+
+class Topic extends Component
+{
+    public function render()
+    {
+        return view('livewire.budgetplan.topic');
+    }
+}
