+++ app/Livewire/CreateAntrag/NewOrganisation.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Livewire\CreateAntrag;
+
+use Livewire\Component;
+
+class NewOrganisation extends Component
+{
+    public function render()
+    {
+        return view('livewire.create-antrag.new-organisation');
+    }
+}
