--- app/Livewire/Forms/ApplicantForm.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Livewire\Forms;
-
-use Livewire\Attributes\Validate;
-use Livewire\Form;
-
-class ApplicantForm extends Form
-{
-    //
-}
