+++ app/Livewire/Forms/ProjectBudgetForm.php
@@ -2,10 +2,11 @@
 
 namespace App\Livewire\Forms;
 
-use Livewire\Attributes\Validate;
 use Livewire\Form;
 
 class ProjectBudgetForm extends Form
 {
-    //
+    public $topics = [];
+
+    public $positions = [];
 }
