+++ resources/views/components/antrag/button/light.blade.php
@@ -1,0 +1,1 @@
+<button type="button" {{ $attributes->merge() }} class="inline-flex justify-center px-3 py-2 text-sm font-semibold leading-6 text-gray-900">{{ $slot }}</button>
