+++ resources/views/components/antrag/button/primary.blade.php
@@ -1,0 +1,5 @@
+<button type="button" {{ $attributes->merge([
+    'class' => "inline-flex justify-center px-3 py-2 text-sm font-semibold text-white bg-indigo-600 rounded-md shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
+])}}>
+    {{ $slot }}
+</button>
