+++ resources/views/components/progressbar/index.blade.php
@@ -1,4 +1,7 @@
-@props([])
+@props([
+    'steps' => null,
+    'active_step' => null,
+])
 <div class="w-full row">
     <nav aria-label="Progress" class="">
         <ol role="list" class="flex items-center grow">
